вынес компоненты в отдельные модули
