+++ .gitignore
@@ -1,2 +1,12 @@
-/.idea/
-/node_modules/
+.DS_Store
+node_modules/
+dist/
+npm-debug.log
+yarn-error.log
+
+# Editor directories and files
+.idea
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
