+++ src/goods-search.vue
@@ -1,0 +1,12 @@
+<template>
+  <form id="searchForm" action="">
+    <input type="search" class="search-button" @input="$emit('input', $event.target.value)"  >
+  </form>
+</template>
+<script>
+  export default{
+      data:()=>({
+          searchLine: '',
+      }),
+  }
+</script>
